[FIX] Table: Correctly create new rows after resize
Recently a row pooling mechanism was introduced to improve the
performance during size adaptations where the row count of the table
changes. Instead of always destroying and recreating all rows, they are
recycled.
Favored by another optimization, where rows are created lazily
only when a binding is available, rows which should be recreated after
invalidation would be recycled from the pool.
With this change it is ensured that the rows in the pool are destroyed
together with the rows in the aggregation.

Change-Id: I54a5986a3b1a73dc7ea5be0edc89339f2a5e5e58
BCP: 1880678505
BCP: 1880674567
BCP: 1880721476
BCP: 1880723120
BCP: 1880722157