[FEATURE] core/routing: support routing in nested components
Up to now routing targets could only be Views. This had the
disadvantage, that Components which are embedded in an app
could not come with their own routing in a configurable way.

This change allows to specify Components as Targets in the
routing configuration of the Component, the manifest.json.
This enables reusable Components to come with their own
routing, which then can be embedded inside any other
Component.

See the Router documentation for details. Routes have been
enhanced to receive an object instead of only a route name
as parameter. There is a new option called prefix. That prefix
serves as identifier for the according Component Target.

JIRA: CPOUIFPHOENIXCORE-2655
Change-Id: Idcb6f4ba8ab2ecd472b6d06b2a13706c28b6667a