[FIX] Demo Kit: Samples page Links don't filter right on first click
Before: On the Main Samples Page in demokit when user clicks on some 
item, its title is being added in the searchbar field and if the title 
contains some whitespace it appears with symbol '%20' instead of ' '. 
The same problem appears in the searchTitle, when navigating to 
search page.

Now: When user selects title with a whitespace inside, it appears in 
the searchbar field as ' '.

Root Cause: After taking title value from the item and adding it to the 
search field, 'hashchanged' event is being fired on window and then the 
hash has been changed in URL encoded string, and the searchfield is 
being updated with the same string. It is fixed by decoding the string 
first.

Change-Id: I41eec2f5f2a5783bc20618f3923f7d1f28c31b5d
BCP: 1880702609