[INTERNAL] Core: fix timing of jquery.sap.dom test
THe jquery.sap.dom test used a 'beforeBootstrap' module to load and
inject some HTML fixture. After injecting the HTML, the module started
QUnit even when the tests had not been loaded yet.

With this change, the test itself loads and injects the HTML (async),
then creates the tests and then lets the starter start QUnit.

Change-Id: Ic53acb2033c62e3fe4f2435d5484e301763a328a