[INTERNAL] sap.m.PlanningCalendar - fix control dev sample.
The toolbar didn't follow the accessibility guidelines for 2 of its
elements - the sap.m.Select(s) at the end were not labelled, but tooltip
was used instead.

This was causing an issue when AT read out the control.

The solution is to have label for each and every select. In addition
making label and select part of the same layout group help them float
together inside the overflowtoolbar.

Change-Id: I8d7b42045e774ebfdf779dbdb52d7004667df662
BCP: 1880701045