[FIX] sap.ui.unified.Calendar - month & year picker are now fully accessible
When Calendar has more than one month, the second set of pickers (above
the second month) were not accessible via keyboard, although its date
cells are accessible via arrow keys.

Providing the second pickers part of the tab chain would be inconsistent
with the cases where 3 or more months are rendered - there is only one
set of pickers(one for month, another for year) for the whole calendar.

Instead, the keyboard recommendation is to always use F4 to open Month
Picker and/or  Shift+F4 to open the Year Picker regardless which is the
currently focused element in the Calendar.

This new key combination should be propagated to the visual impaired
persons, that's why the aria labels for corresponding buttons are enriched.

Change-Id: Ib112b2c96589f662cd5d20cda10c9ffdf9e317ba
BCP: 1870482172