[FIX] sap.ui.unified.ColorPicker: slider texts adjusted
Having both "X slider" and "Slider handle" texts resulted in
unnecessary announcement of the word "slider". This was not helped
by the fact that JAWS also announced "slider" at the end due to
the role="slider". That is why it was decided to rename "X slider"
to "X control".
Note: X = Alpha or Hue
BCP: 1880701326

Change-Id: Ie3e4f9c2f45774cb330f98ebde4e5c4a5891407f