[FIX] sap.m.Popover: Exception when popover opened after being destroyed is fixed
The Popover introduces debouncing when it's openning the Popoup.
That's because of the animations and to deal with a finished state:
open or closed.
There's an edge use case where the destruction comes just before the
Popup's open method is called.
The whole problem could be seen in the qUnit.
It's somehow exploiting the control, but still we should protect it.

Change-Id: I9826010f1e686783ac1aa69eeb2feea94ef77cc7
BCP: 1770249152
Fixes: https://github.com/SAP/openui5/issues/1473
CR-Id: 002075125900004539612018