[FIX] sap.uxap.ObjectPageLayout: Flickering is fixed
Problem: In some cases, when the header is getting collapsed and
unsnapping/snapping measurements are taken, Scrollbar appears and
changes the width of the content, which results in one of Table's
columns title dropping on two lines. This triggers the ResizeHandler
and starts an infinite loop of resizing.

Solutions: Adjust spacer's height in such a way that the Scrollbar does
not appear when it is not needed - so that the content width does not
get changed.

Change-Id: Ie6528917d20fc5374cf59061bf25116e025ce143
BCP: 1870466266