[FIX] sap.m.OverflowToolbar: Fixed shrinking after destroyContent
Problem:
When new content is added after the old content is destroyed, new items
don't shrink. This happens because controls collections are getting
cleared with delay and at this moment the new controls have already been
added to the content aggregation.

Fix:
Clear all controls collections synchronously at the same time when the
content aggregation is destroyed. This allows controls collections to be
filled again with new controls correctly.

Change-Id: If8c8986f3b6bf0355bac2c28e50b0ac10a40b46f
BCP: 1880696108