[FIX] Demo Kit: Header menu items are now fully displayed on mobile
Before: When Demo Kit open on mobile, value of the Select box inside
the header was cut at the end if 'Reference API' or 'Documentation'
option was chosen.

Now: The whole option text is shown in the Demo Kit header on all
devices.

Rootcause: 'Change version' button must be hidden on mobile by default.
It was shown accidentally by missing call of the relative method,
after viewport size is calculated and updated in the model.
Fixed with method call added.

BCP: 1870516107

Change-Id: If0c40ef5e95ec19d3fffdff06582e5f7b021b378