[FIX] sap.uxap.ObjectPageSubSection: Enabled fitContent for BlockBase
Background:
Applications can set the sapUxAPObjectPageSubSectionFitContainer
class to any sap.uxap.ObjetPageSubSection to make it auto-expand
to fit the height of the sections container.

Problem:
The auto-expansion failed if the application content was inside
an instance of sap.uxap.BlockBase.
It failed because the wrapping div of the BlockBase was not
automatically given 100% height (when its parent
sap.uxap.ObjectPageSubSection has the class
sapUxAPObjectPageSubSectionFitContainer)

Solution:
Added the missing css that gives 100% height to the sap.uxap.BlockBase
wrapping div (when its parent sap.uxap.ObjetPageSubSection has the
sapUxAPObjectPageSubSectionFitContainer class)

BCP: 1870528995

Change-Id: Id7337e447ef598f5294415e2f8cec0694dfe6e21