[INTERNAL] check isInvalidateSuppressed
To avoid multiple rendering of the same control during 
databinding we introduced the invalidateSuppressed counter 
some years ago.

When a child control is added to an aggregation, the setParent 
call of the child control could lead to an invalidation. Both, 
the parent as well as the child control registers for rerendering. 
Normally this is no problem as we detect that the child gets already 
rendered by the parent.

If the parent control suppresses rendering for this aggregation 
the parent is not registered for rerendering. We cannot detect that 
the child gets rendered by the parent. That's what the parent check 
in isInvalidateSuppressed is for. The child does not invalidate as 
well and gets rendered by the parent.

This is working fine for aggregations, but not for properties. The
parent does not know when to rerender a child control. So we 
removed this check for property update.


Change-Id: Iffc12234752ef1c470c47c2aae5190842b133a4e