[FIX] Demo Kit: Deep linking to search page results is fixed
Before: When deep linking to a search result with a keyword,
containing a space character and after that refresh the page, this
result was a page with '0' pages found,
when actually search results exist for the given keyword.

Rootcause: In the Search Page controller in the _onTopicMatched,
the parameter, passed later for displaying of the results
(taken from event`s arguments from the URL)
hasn`t been decoded on first place, before being sent to the search API,
and the space inside the query string appears encoded incorrectly.
(%2520)

Fix: DecodeURIComponent method needs to be called on assignment of the
query string variable in order to then pass it in the correct format to
the method reverting this action (EncodeURIComponent).

Change-Id: Iabc08912128e4dce6a50a092fbe1ec1af95c9190
BCP:1870492994