[FEATURE] sap.m.Carousel: Ability to display several items at once
The sap.m.Carousel is now enabled to display several items at once.
This functionality is implemented through a new customLayout aggregation
which is of type CarouselLayoutBase. It defines how many items will be shown
in the visible area of the Carousel control. It has visiblePagesCount
property which determines the default count of items to be shown.
New Sample with customLayoutData aggregation is added to Demo Kit.
Test page of Carousel is also updated in order the new functionality to
be tested.

Change-Id: I9fcb101ca9f0b4595f30dd3ce666994d3426dd17
JIRA: BGSOFUIPIRIN-3337