[FIX] sap.m.MessagePopover: MessageView's auto binding fix
The MessageView has a feature which is when there's no items' binding,
it gets auto bound to sap.ui.getCore().getMessageManager().getMessages().
There are certain checks in the MessageView which are violated when it
gets wrapped into the MessagePopover.
So, the fix is to overwrite that check with the MessagePopover context
when MessageView is used inside.

BCP: 1870298128

Change-Id: I8aa38a72442e17a3f76439bac2f2435d3db7156b
CR-Id: 002075125900004934892018