[FIX] core.support.plugins.ControlTree: context for OData v4 fixed
The Ctrl-Alt-Shift-S popup failed to display the binding context for
controls bound against an OData v4 model. Reason: model.getObject() is
called, but unsupported in that model.
Solution: do not attempt to access data; display the binding context
path in orange as "unverified". This is to avoid showing red or black
when the actual state is not known.

This is related to change ID I9e01b2637c52d152e4fbd6e615aa3418b964d26f

Change-Id: I218175f077659ed6c651e6fb17ce6a2951b03c93