[INTERNAL] testsuite app: align unit testsuites with CI testsuites
The testsuites used in the SAP internal CI have a different structure
and content than the ones in the testsuite app. The latter have to be
aligned in order not to miss any content in the local setup.

Testsuites for codeeditor, fl, rta and uxap had been missing; those for
support and SDK had been added to a different top-level testsuite; the
name of the SDK testsuite was wrong and the new top-level testsuite for
demoapps was completely missing.

Change-Id: I013f230428893318a95f9c325062675fa4886ca6