[FIX] XMLComposite: detect parser error in given fragment content
All browsers behave differently when the fragment content of an
XMLComposite is given directly as string and the DOMParser is used to
parse it toXML: IE11 throws an exception, Firefox logs an error, Chrome
does nothing.
This change makes Chrome also throw an exception, including the message
which has been injected into the parsed XML in a new <parsererror> tag.

BCP: 1970002581

Change-Id: If922c6ad4b7380daf6c9c2cda24eef80cfbbbf1e