[INTERNAL] Fix handling of examples/references when exporting 'api.json'
For symbols other than classes, the examples {@example tags) and
references (@see tags) had not been exported to the api.json.

While fixing this, it was considered to limit the export of other
attributes like 'abstract', 'final', 'extends' and 'implements' to
certain kinds of symbols (mainly to classes). But for each new
limitation, a counter example was found. See TODO comments for details.

Change-Id: I3b4c1f9eb6969747e49f6ac36c535949f3f0fa5c