[FEATURE] sap.f.Avatar: Fallback image behavior added
Until now, when trying to load image from invalid source path, the
Avatar was displaying a blank space inside the placeholder with no
indication that something went wrong. Right now we are handling this
situation by drawing initials type of Avatar if initials are provided
by the user, if not - Avatar with default Icon, and on top of them-
the one with the image. If image doesn`t load we simply don`t see it,
but see the 'fallback' element instead.

Jira BGSOFUIPIRIN-3508

Change-Id: I933d7de29fb5b2cf1c8783b1d064e7bbcc6181ce