[FIX] sap.m.NavContainer: autoFocus false negative fixed
Problem: The app developer is trying to manipulate the focus by using
our pseudo event "onAfterShow", but the focus behavior is not as
expected.

Root Cause: This is not possible if we have focused element
inside the From Page.
After _afterTransitionCallback is called we trigger the AfterShow event
in which the app developer can set the focus on his desired element, but
later on in _applyAutoFocus we are overwriting the focus.
This is caused by the fact that we are checking if the focus is inside
the From Page and if it is we are applying the focus to the last
remembered focused element on the Current Page. We are doing this, even
if the autoFocus property is false.

Solution: We are not calling the _applyAutoFocus function at all, if
autoFocus property is false.
Responsible QUnit tests are corrected respectively.

Change-Id: I974479685499fab9288c0eeb4b92326854f42f98
CR-Id: 002075125800002971922019
BCP: 1870488179
Fixes: https://github.com/SAP/openui5/issues/2306
(cherry picked from commit 6d46cf0b46b9660a6028a51f36f41a3086d15090)