[INTERNAL] sap.ui.support: RuleSetLoader XHR enhancement
Before: RuleSetLoader was making a *sync* HXR request in order to decide
whether internal rules can be loaded or not, for optimization purposes.
This was causing issues with test framework where sync requests are
not supported properly.

After: Code amended in order to provide *async* along with the *sync* API
allowing checking internal rules availability. This allows to make
a single async call to check for internal rules and from then on
to re-use the result in both the sync and async API methods for that purpose.

Change-Id: Ieb734d0e4573f1c1192f07200087a79843262b41