[INTERNAL][FIX] ui5loader: workaround for Safari 12 regression
According to unit test results, Safari 12 seems to optimize write access
to a property with a setter in way that 'freezes' the current setter. If
the setter is changed at a later point in time (as potentially done by
the AppCacheBuster when it is activated), then the optimized code seems
to continue to use the old setter. At least, the new setter is not
called from code that was executed multiple times before.

This change modifies one such write access (setting the src of a script
tag in the ui5loader) so that it is no longer optimized in the way
described way.

If the sketched optimization problem indeed exists, it might occur in
other contexts as well.

Change-Id: I0041871b71488547aca6459c6b8bb04a9a962845
BCP: 1970035485