[FIX] Table: Avoid duplicate id issues for rows
During table destruction, rows would be recreated while unbinding the
rows aggregation if showNoData=false, and stored in the table's row
pool, leaving them in memory.
When navigating to another app and back, a table with the same ID will
be created, which will then try to create rows with IDs that are equal
to the IDs of rows that are still in memory and not destroyed.

Change-Id: I124a649804aeed3e9a5666bdafd3cd59f37ab486
BCP: 0020751294 0000024529 2019