[FIX] m.Link: Self-reference application enhanced
Until now, the Link always added a self-reference, when it had
something in the aria-labelledby attribute in the DOM (via Label,
ariaLabelledBy associtiation, etc...). This mechanism was implemented
because when an element has aria-labelledby, the screen reader will
read it, rather than the element's content. This resulted in the link's
text not announced at all. For IE however, this mechanism was not needed
and only resulted in the text being announced twice. To handle this, we
now check if the browser is IE and apply self-reference only if it's not.

Change-Id: Ic1874653704f452a44f8f1b73a70a5bdd3811e47
BCP: 1870015902