[FIX] sap.m.Image: Fixed failing qUnit tests in IE and Edge
1.Failing in FireFox:

- Root Cause:
There value "auto" is not supported in FireFox for "backgroundSize".

- Solution:
It is changed to "initial" when we are on FireFox.

2.Failing tests in IE and Edge:

- Root Cause:
If the value "center" for "backgroundPosition" exists in a shorthand
with a combination with "right" or "left", it takes effect, but
it is not kept in the Dom Ref.
This is a browser limitation which can be reproduced in the newly
created Snippix "IE Image Limitation" with id "26577".

- Solution:
From "right center" it is changed to "right top".

3.Failing tests in IE only:

- Root Cause ("backgroundPosition"):
The value "initial" is not valid for the "backgroundPosition" property.

- Solution ("backgroundPosition"):
It is changed to "0% 0%", when we are on Internet Explorer.

- Root Cause ("backgroundSize"):
The value "initial" is not valid for the "backgroundSize" property.

- Solution ("backgroundSize"):
It is changed to "0% 0%", when we are on Internet Explorer.
We can't set it to the default value of the browser ("auto") since
it will break the tests in FireFox.

- Root Cause ("backgroundRepeat"):
The value "initial" is not valid for the "backgroundRepeat" property.

- Solution ("backgroundRepeat"):
It is changed to "repeat", when we are on Internet Explorer.

Change-Id: If1c81aa25b3633369485607f6bc05e1469b3580b
BCP: 1970042795