[FIX] sap.m.Button: getCustomData function is overwritten inside CombineButtons
- This fix overwrites the Button getCustomData function in order
to return either the CustomData of the Button or the CustomData
of the MenuItem. This is needed since on combine action, the custom
data is copied from the Button to the MenuItem and inside the
application the app developer tries to get this custom data from the
Button but it is already in the MenuItem.

BCP: 1970043812

Change-Id: I42a98fb38a4e86d142988945988be7cf2c4d39ae
(cherry picked from commit 638bb1dc14915947cc3d5eff066cb8035f1126f0)
CR-Id: 002075125800000258702019