[INTERNAL] Introduce namespaces that are ignored in XMLTemplateProcessor
This change allows to introduce attributes which are only relevant
for XML, e.g. in preprocessing or in support tools, but should not use
browser memory at all and are not needed in the control tree.

Use case:
To be able to unittest expression binding in sap.fe templates and mdc
fragments, we need to mark the elements to be tested. Examples are here
  I56b0ac2f97f4729194aca1b5c5c71b4bb897400f

Change-Id: I58944e9f592cc3d8978590e8d6bd8ce17d5106e6