[FEATURE] sap.f.SemanticPage: Implemented single-line title on phone
Description:

The only content that is displayed in the titleContent
when it is viewed on a phone mobile device and the headerContent
is in collapsed (snapped) state.

Using this aggregation enables you to provide a simple, single-line
title that takes less space on the smaller phone screens when the
headerContent is in its collapsed (snapped) state.

Change-Id: I2777eef2908136350beda6b75d482f0dac583bd3
JIRA: BGSOFUIPIRIN-3513