[INTERNAL] ManagedObjectModel: Handle array/properties of managed object
Note:
Before having this change properties that have been created as arrays
or objects could be used in a binding, via /objectArray/0/value/0 see
unit test. But in contrast to changing the property 'objectArray' of the
managed object inhabiting this property only the JSON part of that value
was changed and so neither an observer nor other bindings where informed.
The purpose of this change is to repair this gap via collecting a node
stack and setting the property with the setter.

Note 2:
For changes we suggest to use the byValue flag.

Note 3:
The unit test shows especially the situation for a test object with
an objectArray property and a binding to a part here.

Change-Id: Iadd5504abe19b2f54e9a9d7aa7e3176658e8bb45