[INTERNAL] sap.f.SemanticPage: Content aggregation is not ignored now
Problem: Due to recent rendering changes in the core, the RTA overlays
for content aggregation of the sap.f.SemanticPage are never created
because of timing issues - the domRef is available at a later time.
This is due to the check in fnShouldIgnoreSingleAggregation, which
requires an already available domRef for a single aggregation to not
being ignored.

Solution: As the actions in the ShareMenu are the only single
aggregations, which may not have a domRef before the ActionSheet menu is
opened, now we do not reuse fnShouldIgnoreSingleAggregation to check if
the content aggregation should be ignored or not. The only check for it
is whether there is such aggregation set or there is not.

Change-Id: If2b4efda977e038b240e7d7b980def7fb5e90923
JIRA: BGSOFUIPIRIN-3660