[FEATURE] ManagedObject: Support handling of properties by value
In recent control development, problems showed up with array / object
properties in ManagedObjects. The values for those properties have not
been treated as immutable, but the current implementation of
ManagedObject requires this to properly detect value changes.

To fix this, a (deep) clone of the values needs to be introduced as a
new option. Managed properties that are declared with 'byValue:true'
will deeply clone their values on write and read access as well as
during a clone of the whole ManagedObject. This decouples the internal
state of such a property from any updates that are made outside.

The clone operation for the values is implemented by the new
sap/base/util/deepClone module. Its semantics are aligned with the
existing deepEqual module so that any clone equals the source.

The 'byValue' option should only be used where absolutely needed as the
additional clone operation might have an impact on performance when used
heavily.

Change-Id: Ibab4b8f39d589772d57cb23a8d64c95b73f8ecfe