[FIX] DateInterval: Check if start date is before end date
Yet, before parsing a given interval string it has not been checked
if the start date is before the end date.

Depending on the strictParsing mode, if it is set to true
the dates get compared now.

In case of an invalid interval string the parsing will "fail"
and return an array with two null values.

Change-Id: I251d2b6fe11de3cd1f9556af901dd13306767a7b
BCP: 1970072645