[FIX] JSONModel: sorting a binding of a flat number array fails on zeros
When a JSONListBinding is created for an array of numbers, sorting the
list binding fails to properly sort zero values in that array. THe same
is true for empty strings in a flat array of strings or for any other
falsy value.

Reason: in the internal lookup method _getObject, falsy values had been
misinterpreted as non-existing context.

Change-Id: I50d7613928735b8f86ff97a2974d42bf97ffbfd5
Fixes: https://github.com/SAP/openui5/issues/2394
BCP: 1970084437