[FIX] sap.m.MultiInput: Deregister resize handler while resizing
MultiInput's _onResize handler specifies some dimensions for the tokenizer.
As MultiInput's default width is 100%, this could lead to resize of the
entire control. This triggers the resize handler.

There are edge cases with certain combination of controls where
this resizing could become an infinite loop: Page/Dialog with scorll
and MultiInput inside. When the MultiInput renders, it could lead to
scroll enablement on the Page, which would result to another resize
in multiinput, which would hide the scroll of the page and so on.

BCP: 0020751294 0000036845 2019
BCP: 1970021491

Change-Id: I2798734ca7e862a6fe3d113d15928085111eea88