[INTERNAL] sap.m.Table - Fiori 3: Default vertical row alignment unified
The default vertical row alignment is now "middle" in the Fiori 3 theme.
This alignment is independent of the selected content density. The
alignment in the belize and belize plus themes remain untouched.

The current belize_hcb and belize_hcw themes behave like the belize and
belize plus theme.

Change-Id: Ife9c8383a5fb539a78a16fe19522ce9e0be75372
JIRA: CPOUIFTEAMB-1274