[INTERNAL] v2/ODataModel: Introduce canonical flag
With this change the new OData model V2 flag 
"canoncialRequets" is introduced.
When this flag is set, the model always try to 
caculate a shorter canonical path of the requested 
resource, before sending a request.

This is useful for OData services, which can only 
handle canonical requests,and no deep requests with 
more than two path segments.

Even if such services are used, OData bindings with 
deep paths can now be used:
e.g.: "SalesOrder(123)/ToLineItems(345)/ToProduct/ToSupplier". 
But this deeppath can only be resolved correctly, 
when the related resources are loaded already (In 
the example above: "SalesOrderItem(345)/ToProduct").

This change is a precondition for an improved OData 
error message handling, which will be introduced in 
a separate change. Since the deep path information,
which can be handed over via the deep path bindings 
to the model, is attached to request objects. This 
additional information could be used later to
create more detailed error messages.

Additionally, a canoncial path cache for the OData 
model is introduced. This cache also improves the 
performance of ODataModel#resolve calls.

Change-Id: I182234c7265c58258b87f13178788bd0f9ede2e4