[INTERNAL] Demokit: removed the misleading version number "v2.0"
To the user the version number of the demokit is confusing because it
conflicts with the actual UI5 version number. Furthermore, this "v2.0"
version is anyway not maintained: since 2.0 of the demokit many
improvements have been added. And the demokit is not a standalone
product where a specific version can be downloaded, but in fact it is
always in sync and bundled with UI5. The term "v2.0" did it's job well
to denote the overhauled Demokit, but memories of the old one are faint.
So all in all, the only significant and maintained version number is the
one of UI5 itself.

Change-Id: Ice5eb58b4dca954d536d589f7259c1b9a56a11d3