[FIX] sap.f.FlexibleColumnLayout: Fixed back navigation in sample
The problem: Back navigation does not update middle column correctly.

The root cause: Navigation to the end column does not store information
about the items in the middle column.

The solution: Update the routing pattern for the last column so that
we know the ID of the item in the middle column.

Change-Id: I0c64455673d71c07417161ac3f84d5f91a88ea5e
BCP: 1870465989