[FEATURE] sap.ui.core: integrated es6-shim to allow restricted ES6 API
Integrates the es6-shim as es6-shim-nopromise as the promise polyfill
isn't compliant to our promise polyfill. The fixes we have done in the
promise polyfill are not available in the es6-shim.

Also the Annex B HTML methods have been removed from the es6-shim-nopromise
as they are also not relevant for the ES6 API compatibility.

For the moment we include the es6-shim-nopromise in the bundles produced
by the build like the sap-ui-core.js or sap-ui-core-nojQuery.js, ...
This increases the size of the bundles but it ensures compatibility rather
than introducing a new and complex loading mechanism which includes the
polyfills only when needed. This can be optimized later.

Change-Id: Id99f3e56206f43fefc72791f9d4150300108b480