[FIX] sap.ui.unified.CalendarOneMonthInterval: Now arrow navigation is correct
This change fixes two issues:

1) If in a PlanningCalendar (1 Month view) Jan 31st was focused, for example,
press on the right arrow resulted in focus on March 1st. Now it focuses Feb 1st
and that's applicable to all months which are following a month with 31 days.

This is caused because of "setUTCMonth" and JavaScript's understanding regarding
Gregorian calendars with 31 days.
 Example: December 31st is selected. If the user sets the month to November,
 since November 31st does not exists, JavaScript gives December 1st.

2) Again 1 Month view in PlanningCalendar with focused date Jan 31st. Press on
the right arrow will show Feb with its 28 days (depending on the year). Another
click on the right arrow will show March with the same dates as February had.

BCP: 002075129500000617722019
Change-Id: Iab2507b025b7d0ffb36d214a271a75690b83a6d1