[FIX] sap.f.Avatar: Fallback content is hidden when image is transparent
When we have non-icon image source added for the sap.f.Avatar
we always load some fallback content 'beneath' it in case
the image fail to load. In the case when this image is
transparent the fallback content was shown in the transparent areas,
so now we make sure we hide it.

Change-Id: I0bd9dcb60edc6ab009475cbbb24b469ef70aa095
BCP: 1970092270