[FIX] m.ViewSettingsDialog: Dialog's title now updating
The reported issue was that upon navigation to the filter details
page, the old title of VSD's dialog would be read out, instead of the
new one (Filter By: XXXXXX).

This was because the old title was still applied as an
aria-labelledby reference on the dialog's root element and the new one
was just visually changed. To overcome this, we now remove the old
title's reference and add the new one when filter page gets opened.
In addition to this, we remove the new one and apply the old when
navigating in the bar, closing the dialog or resetting the filters,
or clicking the "back" button in the filter details.

Change-Id: I14cf66bf262c68771deb339f862ee9842d6d620e
BCP: 1880701367