[INTERNAL] Control: make dependency to ResizeHandler 'soft'
The Control base class itself has no hard requirement to use the resize
handler. It just wants to ensure that all resize handler registrations
made by subclasses or other code are released at destroy time.

This can be achieved with the probing variant of sap.ui.require. When
the ResizeHandler was not loaded, there should be no registrations.

Change-Id: Iaa02f8bedc9f68a75a3551a1c901e61196c88e9b