[FIX] m.PlanningCalendar: Row tooltips are now shown
The issue was reported due to the PlanningCalendarRow's
tooltip not being shown, despite being set. The cause for
this issue was that the 'tooltip' property wasn't propagated
to the PlanningCalendarRowHeader, which would then propagate
it to the StandardListItem underneath.

Change-Id: Id046e249e6098720284bec63446d9aadb6910cc1
BCP: 0020751295 0000066406 2019
CR-Id: 002075125900000678082019
(cherry picked from commit 713f9811b42425b680549526751a4b2e7569a06d)