[FIX] m.PlanningCalendar: Unnecessary row's role attribute removed
The reported issue was that upon pressing INS+F5 while JAWS was opened,
elements with "unknown window type" appeared in the list. This was caused
by the PlanningCalendarRow, which inherits from the StandardListItem.
One option to fix the issue is to provide an explicit aria-label and the
other is to remove the role attribute. Since we remove the tabindex of
the PlanningCalendarRow, this means that the element is no longer accessible,
so we have chosen to remove the role attribute.

Change-Id: I4ab329fdf2766b92cd644adc8cbba8e3c8631580
BCP: 1880702786
CR-Id: 002075125900000575562019