[INTERNAL] Table: Provide CSS class to ensure fixed default row heights
If the "sapUiTableFixedRowHeights" CSS class is applied, the row heights
cannot be increased by the cell content in any "visibleRowCountMode".
If the "rowHeight" property is not set, the heights will always be the
theme and density based default height. If the "rowHeight" property is
set, the heights will always be this height.

Change-Id: I76ec42d6d57a402e1f148f7214c24800167b9fad
Jira: CPOUIFTEAMB-1279