[FIX] sap.m.MaskInput:pressing key in selected value for Android is fixed.
The issue:
On Android devices with Google Chrome if the MaskInput value is selected
and the end user types-in a char, the value was deleted (correct), but
the user-typed char wasn't shown. Instead what the user gets is an empty
string with cursor at the beginning. Typing a char for the second time
works though.

Root cause:
A bug in special handling for Android + Chrome, which did not expect
this scenario.

Fix:
Any value selection is considered.
The virtual "backspace" logic is reworked as well to allow both deletion
and char entering at the same time (see function _buildKeyboardEventInfo).

Change-Id: I7860fd0f0658111d58ed5f3408cbd9b63b39e24e
BCP: 002075129500000552212019