[INTERNAL][FIX] m.FacetFilter: Acessibility for light version adjusted
FacetFilter now displays only it's text (the default one, or which filters
are selected) and adds a hidden 'Filter' label only for screen readers.
In addition to this, references to that hidden label and the FF's text are
now being added to the ariaLabelledBy association of the summary bar,
rather than explicitly generating an aria-label attribute on the DOM element.

Change-Id: Ia1a5f9b2ae6359079b3bb47988a766e8275989ee
BCP: 1780150816
CR-Id: 002075125900000730252019
(cherry picked from commit a1e117867e7874381adcd00354af6667091d45b0)