[INTERNAL] sap.ui.unified.CalendarOneMonthInterval: Now qunit tests passes
The test was failing because is was using innerText which works differently
on the different browsers. The test's purpose was to compare two texts -
a visible one on the screen and the expected one.
That's why now it'll use textContent, which is not aware of the rendered
appearance of the text instead of the innerText.

Also a minor documentation typo is fixed.

Change-Id: I4501985a0afc885f60576a3ca6c05b6ab0049277
BCP: 1970117151