[INTERNAL] sap.ui.dt: improve logic for scrollbar margins when overlay is selected
The current state:
1. if control has an aggregation with scrolling and this aggregation
takes the whole space of the control, then the selection border is
painted inside the control with margin for the scrollbar width.
On other hand, e.g. in Master View in RTA Freestyle Application, the
control has several aggregations and only one of them has a scrollbar,
then the selection border for this control will be painted around
the control's bounding dimension (control's geometry) with a slight
scrollbar overlapping.

2. if control has scroll containers, the logic is the same as in step 1
with additional check for each scroll container.

BCP: 1970012394
Change-Id: If866548bc3c25c6198c268c0e3bf36eee46a9a26