[INTERNAL] core.Element: don't create new CustomData after destroy
The Element.prototype.data method should not create new CustomData for
an already destroyed element. For sure, no modification makes sense on a
destroyed element, but throwing an Error would be incompatible for
mutators and logging errors would add code but not prevent dead objects.
Method data() is different as it creates new objects on its on and that
can be fully prevented.

Change-Id: Iea9c8382089ebe7e7aa613144e484b9dd516e6e6
BCP: 1870528703