[INTERNAL] DebugMode QUnit Test: Report dummy test
Adds a dummy test to report at least one test result in case the
preconditions are not met.

It seems that currently the voter is not running with the nightly.build
profile which causes no debug modules to be created which is one
precondition.
This will be adressed as a follow up.

(cherry picked from commit e03472988b4acd2d7796d34cc18b06cb8e6cb5f2)
Change-Id: I7674d15a8919897166950e6b3e074ccb6efaab2b
CR-Id: 002075125900000889532019
BCP: 1970131205