[INTERNAL] DebugMode QUnit Test: Report dummy test
Adds a dummy test to report at least one test result in case the
preconditions are not met.

It seems that currently the voter is not running with the nightly.build
profile which causes no debug modules to be created which is one
precondition.
This will be adressed as a follow up.

BCP: 1970131205
Change-Id: I7674d15a8919897166950e6b3e074ccb6efaab2b