[INTERNAL][FIX] m.FacetFilter: Acessibility for light version adjusted
FacetFilter now displays only it's text (the default one, or which filters
are selected) and adds a hidden 'Filter' label only for screen readers.
In addition to this, references to that hidden label and the FF's text are
now being added to the ariaLabelledBy association of the summary bar,
rather than explicitly generating an aria-label attribute on the DOM element.

CR-Id: 002075125900000731702019
BCP: 1780150816
Change-Id: Iab4df5af9478f3e0b584e0a0b2e6bd9b9606aaee