[INTERNAL] ODataMessageParser: Filter out duplicate messages
With this change the outer errors, which are only used as message container for
inner errors, are filtered out also for the "Response Header in JSON format" scenario.
The message containers are detected through the same message properties to a inner error
(relevant properties are status code and message text).

Change-Id: Iec5c1463ffdf0683cec922f12742934dd123abf6