[INTERNAL] sap.f.DynamicPage: Fixed Fiori 3 collapse/pin buttons layout
Problem: Incorrect alignment of the collapse/expand and pin buttons icons.

Root Cause: After the rework of the sap.m.Button, additional alignment
was needed for the collapse/expand and pin buttons in the sap.f.DynamicPage.

Solution: Unnecessary padding of the icons in collapse/expand and pin
buttons is removed.

Change-Id: I7d4acd033acf96f503e02e8398d2a606db95816c