[INTERNAL][FIX] sap.m.Carousel: Offset is correctly calculated
Problem: In some cases, due to animations upon going from one page to
another, the offset of the mobify carousel is calculated too early and
this results in wrong positioning of Carousel's items.

Solution: Calculation of the offset is done in the _update function,
called with a delay (requestAnimationFrame), after all animations are
over.

Change-Id: I65e5d944b8e87dde954a2a423380554a27042dcd
BCP: N/A