[FIX] sap.m.ProgressIndicator: Invalid percentValue handling corrected
BCP: 1970134215

Problem: With an earlier fix we have forbidden providing other types
except numbers for the percentValue property, resulting in an exception
thrown if the provided value is not correct.

Solution: The value now is first casted to the number type, then checked
for the correct range between 0 and 100.
If it's correct the percentValue property is set, otherwise the propety
will not be set to a wrong value and a console warning is present to the
user.

Change-Id: If2615ea76862fed61fae7d75c28d357ae1a57b64