[INTERNAL][FIX] sap.f.ShellBar - Fixed responsiveness on mobile
The problem:
On mobile phone the responsiveness of the control is broken and is not
working as expected. Especially if you switch from portrait to landscape
mode several times the control becomes unusable

The root cause:
The part of the code handling responsiveness was not optimal and needed
improvement

The solution:
Responsiveness of the control refactored and extracted in a class for
more simplified implementation and maintenance ease.

Additional fixes:
* Theme parameters improved
* Code simplification
* Fixed a bug with static area not available during measurement
resulting in a fatal error

BCP: 1970151611
CR-Id: 002075125900000888922019

Change-Id: Ic724fb5c12eb3fe190f12235a89088362b4b6660
(cherry picked from commit c7a4e7a5c20768d27d1a339c6e8bf1b59401bbe5)