[FIX] sap.m.Select: First item in list can now be selected on mobile
Problem: When the user opens the select on a mobile
device the first item in the list cannot be selected.

Root cause: The sap.m.Select uses a sap.m.Dialog to display
the value state message. The sap.m.Dialog has a subHeader
with no content, which appears on top of the first item.

Solution: Set the subHeader's display property to 'none'
when the control's valueState is ValueState.None

Change-Id: I32d1e7b156cf6a0f4a4b6d68717498896b5b6fdb
BCP: 1970141576